move data-fetch request from App file to component file
