+++ src/App.tsx
@@ -1,5 +1,4 @@
-import React, { useEffect, useState } from "react";
-import axios from "axios";
+import React, { useState } from "react";
 
 import UsersTable from "./components/usersTable";
 import AddUserForm from "./components/addUserForm";
@@ -8,36 +7,11 @@ import AddUserForm from "./components/addUserForm";
 import "./App.css";
 
 function App() {
-  const [usersList, setUsersList] = useState([{}]);
-  const [isLoading, setIsLoading] = useState(false);
-  const [error, setError] = useState(null);
   const [isNewUser, setIsNewUser] = useState(false);
-  const [isRemoved, setIsRemoved] = useState(false);
   const [formIsVisible, setFormIsVisible] = useState(false);
-
-  const fetchUsersHandler = async () => {
-    setIsLoading(true);
-    setError(null);
-    try {
-      const response = await axios.get("https://63a19d4fba35b96522e2ff4e.mockapi.io/users");
-      const { data } = response;
-      setUsersList(data);
-    } catch (error: any) {
-      setError(error.message);
-    }
-    setIsLoading(false);
-  };
-
-  useEffect(() => {
-    fetchUsersHandler();
-  }, [isNewUser, isRemoved]);
 
   const confirmHandler = (isSubmitted: boolean) => {
     setIsNewUser(isSubmitted);
-  };
-
-  const deleteHandler = (isDeleted: boolean) => {
-    setIsRemoved(isDeleted);
   };
 
   const showFormHandler = () => {
@@ -46,24 +20,6 @@ function App() {
   const formCancelHandler = () => {
     setFormIsVisible(false);
   };
-
-  let content;
-
-  content = <p>Found no Content!</p>;
-
-  // if (usersList.length) {
-  //   content = (
-  //     <UsersTable usersList={usersList} onDelete={deleteHandler} onConfirm={confirmHandler} />
-  //   );
-  // }
-
-  if (error) {
-    content = <p>{error}</p>;
-  }
-
-  if (isLoading) {
-    content = <p>Loading...</p>;
-  }
 
   return (
     <div className="App">
@@ -76,7 +32,7 @@ function App() {
             Add a User +
           </button>
         </div>
-        <UsersTable usersList={usersList} onDelete={deleteHandler} onConfirm={confirmHandler} />
+        <UsersTable onConfirm={confirmHandler} isNewUser={isNewUser}/>
       </>
     </div>
   );
